requirements.txt
Django
psycopg2
psycopg2-binary
django-oauth-toolkit
djangorestframework
django-cors-headers
django-rest-swagger
django-filter
rest_condition
Pillow
boto3
django-storages
django-multiselectfield
gunicorn
django-cors-headers
pyzbar
pyqrcode
pypng
django-resized
django-cities-light
drf-yasg
packaging
django-push-notifications
django-constance



